Solve timeCache update
The cache would return the old value if referenced from outside of
time.go. Moving it out of the type and into a global variable solves
this, but it feels rather unidiomatic and a proper solution is still to
be found.
An original suspicion was the Goroutine, but adding a Mutex lock around
the store operation did not help.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>